requirements.txt
llama-index
llama-index-llms-gemini
llama-index-embeddings-huggingface
chromadb
streamlit
python-dotenv
biopython
pydantic
openai
